refactor: rename files to follow Radix naming convention
