+++ packages/raystack/v1/components/announcement-bar/index.tsx
@@ -1,0 +1,1 @@
+export { AnnouncementBar } from "./announcement-bar";
