--- packages/raystack/v1/components/announcementBar/index.tsx
@@ -1,1 +1,0 @@
-export { AnnouncementBar } from "./announcementBar";
