+++ packages/raystack/v1/components/dropdown-menu/dropdown-menu.tsx
@@ -2,7 +2,7 @@ import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 import { cva, VariantProps } from "class-variance-authority";
 import * as React from "react";
 
-import styles from "./dropdownMenu.module.css";
+import styles from "./dropdown-menu.module.css";
 
 const content = cva(styles.content);
 const DropdownMenuContent = React.forwardRef<
