--- packages/raystack/v1/components/dropdownMenu/index.ts
@@ -1,1 +1,0 @@
-export { DropdownMenu } from "./dropdownMenu";
