+++ packages/raystack/v1/components/empty-state/empty-state.tsx
@@ -2,7 +2,7 @@ import clsx from "clsx";
 
 import { Flex } from "../flex";
 import { Text } from "../text";
-import styles from "./emptyState.module.css";
+import styles from "./empty-state.module.css";
 
 type classNameKeys =
   | "container"
