+++ packages/raystack/v1/components/empty-state/index.tsx
@@ -1,0 +1,1 @@
+export { EmptyState } from "./empty-state";
