--- packages/raystack/v1/components/emptyState/index.tsx
@@ -1,1 +1,0 @@
-export { EmptyState } from "./emptyState";
