+++ packages/raystack/v1/components/icon-button/index.tsx
@@ -1,0 +1,1 @@
+export { IconButton } from "./icon-button";
