--- packages/raystack/v1/components/iconButton/index.tsx
@@ -1,1 +1,0 @@
-export { IconButton } from "./iconButton";
