--- packages/raystack/v1/components/inputField/index.tsx
@@ -1,1 +1,0 @@
-export { InputField } from "./inputField";
