+++ packages/raystack/v1/components/text-area/index.tsx
@@ -1,3 +1,3 @@
-export { TextArea } from "./textArea";
+export { TextArea } from "./text-area";
 
 // Todo: Add tooltip for the info icon
